voice/flask-voice/requirements.txt
flask
africastalking
python-dotenv
flask-cors
fastapi
uvicorn